refactor: http folders
